+++ src/graphics/shader.cpp
@@ -188,7 +188,7 @@ bool Shader::LoadAtlas(const char* filename)
 		}
 		else if (line_trimmed[0] == '#')
 		{
-			int pos = line_trimmed.find_first_of(' ');
+			int pos = (int)line_trimmed.find_first_of(' ');
 			if (pos != std::string::npos)
 			{
 				std::string cmd = line_trimmed.substr(1, pos - 1);
@@ -220,9 +220,9 @@ bool Shader::LoadAtlas(const char* filename)
 		line = trim(line);
 		if (line.size() == 0 || line.substr(0, 2) == "//")
 			continue;
-		int pos = line.find_first_of(' ');
-		int pos2 = line.find_first_of(' ', pos + 1);
-		int pos3 = line.find_first_of(' ', pos2 + 1);
+		int pos = (int)line.find_first_of(' ');
+		int pos2 = (int)line.find_first_of(' ', pos + 1);
+		int pos3 = (int)line.find_first_of(' ', pos2 + 1);
 		if (pos3 == -1)
 			pos3 = static_cast<int>(std::string::npos);
 		std::string name = line.substr(0, pos);
